Handle auth errors and redirects
There's a bug where we lose the flash message due to a double redirect. This also removes a bunch of duplicate logic that was already in the middleware. The session creator is now "dumb" in that it creates a session and then relies on the middleware to verify authenticity, reducing duplication.