pkg/database: alleviate User CRUD load by a distributed cache
Adds Redis as a pass-through cache for the methods of *Database
for CRUD for users. Write-throughs are performed on Create and Update,
Read-through for Read, and Deletion for Delete.

This change has an escape hatch for disjoint consistency with
that database in that, if a value isn't updated, its value will
expire in 1 day (86,400 seconds)

Also added tests to verify the pass-through by intercepting each
command sent in.

Fixes #30