Separate database config loading from connection opening (#225)
The previous "Open" command would parse the configuration and then establish the connection. This separates into two separate invocations. This is important because retrying Open() was actually unsafe in local development because config parsing mutated state. By moving config parsing to be separate, we can parse once, then call Open() as many times as we'd like.