Add more generic KMS handling
This moves the logic out of the SMSConfig and into the database. It has the added benefit of not needing to store the database reference on the SMS config to access the keys and key manager. It also allows us to trivially add encryption to another column of a table in the future.