Add model-level validation, clarify db relationships, add users CRUD operations (#233)
* Separate database config loading from connection opening

The previous "Open" command would parse the configuration and then establish the connection. This separates into two separate invocations. This is important because retrying Open() was actually unsafe in local development because config parsing mutated state. By moving config parsing to be separate, we can parse once, then call Open() as many times as we'd like.

* Add more generic KMS handling

This moves the logic out of the SMSConfig and into the database. It has the added benefit of not needing to store the database reference on the SMS config to access the keys and key manager. It also allows us to trivially add encryption to another column of a table in the future.

* Add generic error handling

* Fix the realm<->user infinite loop and clarity ownership in ERD

* Add updated users interface

* Small misses from API change

* Review comments