Adding support for App Links, fallback to Play Store, fix regression. (#766)
Redirector will first redirect to itself,
on Android adding bounce=1 to ensure app link has a chance
to trigger (work around for some SMS clients).

iOS is also supported, but for now assuming it doesn't need
the extra bounce.

When an App Link is not configured for a platform,
continue to fall back to ens://

Fixing regression in propagating query parameters.