Give system admins ability to clear caches (#834)
* Give system admins ability to clear caches

In some situations, a system admin may want to clear the caches, either
for development/testing or because a misconfiguration has made it into
the system. This introduces a way for system admins to perform cache
resets.

This required a bit of a refactor of the cacher, separating the
namespace property from the actual key value.

* Update descriptions

* Use public_keys