Add test harness for headless browser testing
After evaluation, I decided to use chromedp since it's a pure-go
solution that should work headlessly on our existing CI infrastructure
(fingers crossed).

There's some open TODOs, but I'm trying to chunk this work since it's
pretty large.

Open questions:

- We need firebase configured to actually login. Which project should we use? How should that config be supplied? For now I hardcoded our firebase config (they aren't a secret), but that's not a good long-term solution, especially if other folks want to run these tests.

This commit also adds a few new internal-only packages to help with testing:

- `internal/project` - project-wide helpers. Currently the only function
is one that returns the filesystem root of the project, which is needed
for telling the server where to find its assets. It will also be used by
enx-redirector and any other filesystem-ey things.

- `internal/envstest` - globally shared test helpers. I'm going to start
centralizing things here across all our test suites. This package also
includes all the logic and helpers for spinning up a full in-memory UI
server (with database and dependencies) for local testing.

- `internal/browser` - wrapper around chromedp. It's small, but I expect
to add more helpers and things as we expand our testing.