Add test harness for headless browser testing (#881)
After evaluation, I decided to use chromedp since it's a pure-go
solution that should work headlessly on our existing CI infrastructure
(fingers crossed).

This also adds a few new internal-only packages to help with testing:

- `internal/project` - project-wide helpers. Currently the only function
is one that returns the filesystem root of the project, which is needed
for telling the server where to find its assets. It will also be used by
enx-redirector and any other filesystem-ey things.

- `internal/envstest` - globally shared test helpers. I'm going to start
centralizing things here across all our test suites. This package also
includes all the logic and helpers for spinning up a full in-memory UI
server (with database and dependencies) for local testing.

- `internal/browser` - wrapper around chromedp. It's small, but I expect
to add more helpers and things as we expand our testing.