Add ExternalIssuers, refactor statistics APIs (#1198)
* Add AuditID issuance path.

Fixes #1076

* Refactor statistics and visualizations

This commit refactors stats and stats APIs to return standardized responses. Instead of munging JSON on the server and the client, the server defines a custom JSON marshaler and the client uses the JSON API to request the chart data. This greatly simplifies the HTML template and controller logic.

In the same vein, there's a custom MarshalCSV interface that all the stats providers implement, and new renderer functions for creating CSVs.

The biggest part of this refactor is the way stats are calculated. Previously stats were naively calculated with a JOIN, but that left zeros or missing data in the structure. The new stats calculation ensures no gaps and fills in zero values where data is missing.

* Review comments

Co-authored-by: Jeremy Faller <jeremy@golang.org>