Ping db on health instead of opening new connection (#1503)
The previous logic relied on creating a new database connection. This
would be a different database connection from the one the rest of the
app is using, and could lead to false successes. This changes the health
endpoint to use the same db that the other handlers use.

Additionally, since it's just a ping now (instead of a full config
parsing), the limit is increased from 1/min to 1/sec. The response code
when the limit is exceeded is now a 429 instead of silently falling
through to 200.

Finally, the monitoring uptime checks are updated to use this endpoint
instead of the generic /health. It's safe to say that the service is
"down" if it cannot connect to the database.