Unify in-memory server bootstrap
This brings the API server and Admin API server bootstrapping into
envstest like the UI server, making it possible to spin up in-memory
instances of both. This removed a TON of duplication between envstest
and testsuite, eliminating the need for some intermediate packages.

This also unifies the 3 different "client library" packages floating
around into a single internal/clients package that supports the superset
of all the APIs, dangled off an appropriate API client.