Check session expiration on login (#1549)
This fixes an issue where a realm has configured session expiration, but the user lands on the login page. Prior to this commit, the login page would re-authenticate, even if the time since last activity exceeded the configuration. This forces the login controller to check the session TTL before rendering the HTML (which would refresh the TTL).