Add rotation harness and rotate token signing keys
This adds a new service, rotation, which handles key and rotation events. It's designed to be invoked via Cloud Scheduler and has its own locking and time intervals. It's conceptually similar to the cleanup job.

This also adds the database schema and example rotation for token signing keys. Note that this code is NOT in use in main code paths yet.