Introduce configurable Authenticated SMS failure modes (#1834)
This introduces a runtime configurable to control how the system behaves when signing an authenticated SMS fails. The default behavior is to "fail open" (continue on error), meaning if the signing fails for any reason, the system still continues and delivers the SMS. Operators can also configure the system to "fail closed" (fail on error) to halt and NOT send SMS unless signing succeeds. To control the behavior, operators must set the `SMS_FAIL_CLOSED` environment variable on the `server` and `adminapi` services. Set the value to `true` to fail closed. The default value is false (fail open).

Regardless of the configuration, an error is logged and a metric is incremented when an SMS fails to sign. There is also a new alerting policy for when more than 5 SMS messages fail to sign in a 60s window by realm. This ensures that, even if the system is configured to fail open, operators are aware of the issue and can respond accordingly. The alert is non-paging since there is rarely action an operator can take if the upstream key management system is unavailable.