Send authenticated SMS in e2e-runner
This updates the e2e-runner to send an SMS credential as part of its test. This feature is on by default, but it can be disabled by setting `E2E_SKIP_SMS` to any truthy value on the e2e-runner service.

Operators must configure the e2e realm with [Twilio Test Credentials](https://www.twilio.com/docs/iam/test-credentials) via the verification server UI. Note that the system will not actually send an SMS, it uses Twilio's magic numbers and test credentials to simulate all aspects of the system and API calls but stops short of actually delivering an SMS.