Cleanup e2e API keys that have existed for > 24 hours
The e2e-runner does self cleanup, but sometimes the job times out or fails. In those cases, there's no opportunity for cleanup and we gradually leak API keys over time.

This introduces a function for getting the e2e-realm (if one exists), and adds a new task to the cleanup job to delete e2e keys that have existed for more than 24 hours.