Track API key "last used" (#2022)
This isn't an exact timestamp tracker, since we use some pretty aggressive caching to avoid a write on every read. However, this will give realm admins a very clear signal of whether an API key is still being used.

The UI is hidden behind a feature flag because I want to capture this data for a full release until displaying it in the UI. Otherwise it could lead to misleading information about an API having "never" been used.