Update CSS/JS elements to their latest versions
The new version of firebase and firebase-auth require a significant refactoring of our Javascript code; we would need to migrate to modules and use a bundler like Webpack. Fortunately there's a compat layer we can use for now to minimize code changes, but this is something we'll need to think about long term.

Fixes GH-2357