Update Terraform dependencies (#2363)
This updates the minimum required Terraform version to 1.2 and updates the Terraform providers to their latest version. As part of this upgrade, some fields were deprecated or renamed.

As part of this change, you will see the local ".env" file deleted and re-created. This is okay because this file is only used for local development.